requirements.txt
accelerate==0.32.1
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorcet==3.1.0
contourpy==1.2.1
cycler==0.12.1
filelock==3.15.4
Flask==3.0.3
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
huggingface-hub==0.23.5
idna==3.7
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
lightning-utilities==0.11.5
MarkupSafe==2.1.5
matplotlib==3.9.1
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pillow==10.4.0
psutil==6.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.3.3
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
setuptools==70.3.0
six==1.16.0
sympy==1.13.0
tbb==2021.13.0
tokenizers==0.19.1
torch==2.3.1
torchmetrics==1.4.0.post0
tqdm==4.66.4
transformers==4.42.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
yarl==1.9.4
